trae - temporary.
file downloads in inbounds successfully(idk if its searching for
message again which is a bad method, or if it is downloading
using the message id found in metadata(that would be right method)
-> correct this if this is not using the correct method.

but file downloads successfully but in inbounds dbfs.

problems:
still not downloading in downloads folder. some error is coming