UI changes: study these changes. beautiful changes. (temp. chkpt.)
1. The header bar has an animation like something is flowing
through it. like waves in the ocean.

(Ocean is our brand identity because its blue. and we want our
product to represent that we're the most safest technology of
storage solutions on the planet.)

2. The header icon has white icon, it grows with size, you may
wanna know how to do these animations in HTML/CSS somewhat.

So you need to study this commit.