fixes #22
1. I don't why but the previous commit changed something in
ChatGroupSelector.js rendering of whose in TelegramMessenger.js
was causing the slowdown of whole creation of folders and
everything, maybe because it is also using the telegramClient
and interfering with it trying to do the same things we're
doing elsewhere.

2. We didn't really needed to render ChatGroupSelector.js,
so I disabled it with JSX comment.

3. Note: If someday we need to debug our code, we'd need to see
this TelegramMessenger.js which was our primitive version
where a small section on the bottom left used to appear to
select group. We'd need to someday, then we'd need to enable it
from here.

Note: 2. Also note, the contents of ChatGroupSelector.js are
changed, in next commit, I'll revert all those changes to restore
it to previous state, we don't really need any change in these
files as they were working correctly as always, and we don't need
to develop them further either, let them be as they are.