Reverting part of 465fa995756e788d5f38e56c30e7bd92f6460dc0
these were changes that were unnecessarily made.
ChatGroupSelector.js is a module in our legacy TelegramMessenger.js
when we first developed the prototype for selecting the
groups in the left bottom of our screen. This was always
working and we don't really need any change to it. whenever,
later we'll need it we'll enable it in TelegramMessenger.js i.e.
render it as it was working before, thus no real change is
necessary in this place really. so reverting this part of
change.